Replace Vedeu::Models::Cell with Vedeu::Cells::Empty.
