+++ lib/vedeu/models/all.rb
@@ -9,7 +9,6 @@ module Vedeu
 end # Vedeu
 
 require 'vedeu/models/toggleable'
-require 'vedeu/models/cell'
 require 'vedeu/models/focus'
 require 'vedeu/models/row'
 require 'vedeu/models/page'
