+++ test/lib/vedeu/renderers/terminal_test.rb
@@ -13,17 +13,17 @@ module Vedeu
       let(:output)    {
         Vedeu::Models::Page.coerce([
           [
-            Vedeu::Models::Cell.new(position: [1, 1]),
+            Vedeu::Cells::Empty.new(position: [1, 1]),
             Vedeu::Views::Char.new(name: _name, value: 'a', position: [1, 2]),
-            Vedeu::Models::Cell.new(position: [1, 3]),
+            Vedeu::Cells::Empty.new(position: [1, 3]),
           ], [
-            Vedeu::Models::Cell.new(position: [2, 1]),
+            Vedeu::Cells::Empty.new(position: [2, 1]),
             Vedeu::Views::Char.new(name: _name, value: 'b', position: [2, 2]),
-            Vedeu::Models::Cell.new(position: [2, 3]),
+            Vedeu::Cells::Empty.new(position: [2, 3]),
           ], [
-            Vedeu::Models::Cell.new(position: [3, 1]),
+            Vedeu::Cells::Empty.new(position: [3, 1]),
             Vedeu::Views::Char.new(name: _name, value: 'c', position: [3, 2]),
-            Vedeu::Models::Cell.new(position: [3, 3]),
+            Vedeu::Cells::Empty.new(position: [3, 3]),
           ],
         ])
       }
